requirements.txt
pytest==7.1.2
selenium==4.9.0
allure-pytest==2.11.1
ruff==0.0.280